# Minor WireFormatting code cleanups and optimizations.