## Proposed Changes

Code cleanups to simplify/optimize code for WireFormatting.

## Types of Changes

- [ ] Bug fix (non-breaking change which fixes issue #NNNN)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause an observable behavior change in existing systems)
- [ ] Documentation improvements (corrections, new content, etc)
- [ ] Cosmetic change (whitespace, formatting, etc)

## Checklist

- [X] I have read the `CONTRIBUTING.md` document
- [X] I have signed the CA (see https://cla.pivotal.io/sign/rabbitmq)
- [X] All tests pass locally with my changes
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)
- [ ] Any dependent changes have been merged and published in related repositories

## Further Comments

A simple change, performance benchmarks shows considerable improvements to serialization times for empty/null strings and objects. Getting rid of "unsafe" code for the `netcoreapp` target.

**Before:**
``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.20262
Intel Core i7-10700 CPU 2.90GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=5.0.100
  [Host]   : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT
  ShortRun : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT

Job=ShortRun  IterationCount=3  LaunchCount=1  
WarmupCount=3  

```
|                                 Type |                Method |      Mean |     Error |    StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------- |---------------------- |----------:|----------:|----------:|------:|--------:|-------:|------:|------:|----------:|
|         WireFormatting_Read_BasicAck |          ReadFromSpan |  3.905 ns | 0.4919 ns | 0.0270 ns |  1.00 |    0.00 | 0.0038 |     - |     - |      32 B |
|     WireFormatting_Read_BasicDeliver |          ReadFromSpan | 11.831 ns | 1.2428 ns | 0.0681 ns |  3.03 |    0.03 | 0.0067 |     - |     - |      56 B |
|  WireFormatting_Read_BasicProperties |          ReadFromSpan | 63.368 ns | 4.6002 ns | 0.2522 ns | 16.23 |    0.10 | 0.0229 |     - |     - |     192 B |
|     WireFormatting_Read_ChannelClose |          ReadFromSpan |  7.086 ns | 0.9426 ns | 0.0517 ns |  1.81 |    0.03 | 0.0038 |     - |     - |      32 B |
|        WireFormatting_Write_BasicAck |      WriteArgumentsTo |  1.899 ns | 0.0449 ns | 0.0025 ns |  0.49 |    0.00 |      - |     - |     - |         - |
|    WireFormatting_Write_BasicDeliver |      WriteArgumentsTo | 34.764 ns | 3.9156 ns | 0.2146 ns |  8.90 |    0.08 |      - |     - |     - |         - |
| WireFormatting_Write_BasicProperties | WritePropertiesToSpan | 32.445 ns | 0.2154 ns | 0.0118 ns |  8.31 |    0.06 |      - |     - |     - |         - |
|    WireFormatting_Write_ChannelClose |      WriteArgumentsTo | 13.923 ns | 1.5404 ns | 0.0844 ns |  3.57 |    0.03 |      - |     - |     - |         - |

**After:**
``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.20262
Intel Core i7-10700 CPU 2.90GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=5.0.100
  [Host]   : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT
  ShortRun : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT

Job=ShortRun  IterationCount=3  LaunchCount=1  
WarmupCount=3  

```
|                                 Type |                Method |      Mean |      Error |    StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------- |---------------------- |----------:|-----------:|----------:|------:|--------:|-------:|------:|------:|----------:|
|         WireFormatting_Read_BasicAck |          ReadFromSpan |  3.889 ns |  0.5254 ns | 0.0288 ns |  1.00 |    0.00 | 0.0038 |     - |     - |      32 B |
|     WireFormatting_Read_BasicDeliver |          ReadFromSpan | 11.002 ns |  0.4028 ns | 0.0221 ns |  2.83 |    0.02 | 0.0067 |     - |     - |      56 B |
|  WireFormatting_Read_BasicProperties |          ReadFromSpan | 58.720 ns | 11.1702 ns | 0.6123 ns | 15.10 |    0.24 | 0.0229 |     - |     - |     192 B |
|     WireFormatting_Read_ChannelClose |          ReadFromSpan |  6.678 ns |  0.1788 ns | 0.0098 ns |  1.72 |    0.02 | 0.0038 |     - |     - |      32 B |
|        WireFormatting_Write_BasicAck |      WriteArgumentsTo |  1.820 ns |  0.0432 ns | 0.0024 ns |  0.47 |    0.00 |      - |     - |     - |         - |
|    WireFormatting_Write_BasicDeliver |      WriteArgumentsTo |  4.797 ns |  0.3201 ns | 0.0175 ns |  1.23 |    0.00 |      - |     - |     - |         - |
| WireFormatting_Write_BasicProperties | WritePropertiesToSpan | 34.957 ns |  0.8850 ns | 0.0485 ns |  8.99 |    0.06 |      - |     - |     - |         - |
|    WireFormatting_Write_ChannelClose |      WriteArgumentsTo |  3.786 ns |  0.1707 ns | 0.0094 ns |  0.97 |    0.01 |      - |     - |     - |         - |