@bollhals 2020-11-26
I like the change 👍  didn't yet think about it yet now that we have the new APIs available.

One question though: How likely is a null / empty string being passed for real world examples?