@stebet 2020-11-26
> I like the change 👍  didn't yet think about it yet now that we have the new APIs available.
> 
> One question though: How likely is a null / empty string being passed for real world examples?

Likely given that it's a valid Exchange for example (the default exchange) :) also, people initializing arrays or lists, aiming to put stuff into and sometimes omitting data, so I think it's a valid use-case.